# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Atlas",
        "description": "Documentación WPARK - API",
        "contact": {
            "email": "lchirinos@acercapartners.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/blocks": {
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Blocks List",
                "description": "List of blocks",
                "operationId": "indexBlocks",
                "responses": {
                    "200": {
                        "description": "Block list Successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Create New Block",
                "description": "Create New Block",
                "operationId": "storeBlock",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlockStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create New Block"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/blocks/{id}": {
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Show Block Details",
                "description": "Show Block Details",
                "operationId": "showBlock",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Block Details"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Update Block",
                "description": "Update Block",
                "operationId": "updateBlock",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlockUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Block"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Delete Block",
                "description": "Delete Block",
                "operationId": "destroyBlock",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete Block successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Restore Block",
                "description": "Restore Block",
                "operationId": "restoreBlock",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Block restored successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlockStoreRequest": {
                "title": "Block Store Request",
                "description": "Block Store request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del bloque",
                        "type": "string",
                        "maxLength": 255,
                        "example": "BLOQUE ZP"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BlockStoreRequest"
                }
            },
            "BlockUpdateRequest": {
                "title": "Block Update Request",
                "description": "Block Update request body data",
                "required": [
                    "name",
                    "active"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del bloque",
                        "type": "string",
                        "maxLength": 255,
                        "example": "BLOQUE ZP"
                    },
                    "active": {
                        "description": "Indica si el bloque está activo (0: No está activo, 1: Está activo)",
                        "type": "boolean",
                        "maxLength": 1,
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BlockUpdateRequest"
                }
            },
...
        },
        "responses": {
...
            "Forbidden": {
                "description": "Permission Error."
            },
            "InternalServerError": {
                "description": "Internal Server Error"
            },
            "NotFound": {
                "description": "Resource Not Found."
            },
            "Unauthorized": {
                "description": "Unauthorized Error."
            },
            "UnprocessableEntity": {
                "description": "Validation Error."
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "id, eg; 1",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}